Prototyping features
Working on getting the base features and input system working.

For example getting it so you can:
- make a custom template that fits the format.

Basic implementation of reading in the templates

Created a 2nd rust file to currently hold the external functions so that its easier to navigate between the functions and main.

Changed the way that it looks at templates and how they are structured.